R&D/Demo AI/requirements.txt
# JLPT Audio Splitter Requirements
# Version: 2.1 - Minimal dependencies for core functionality

# Core
python-dotenv==1.0.0

# Audio Processing
openai-whisper>=20231117
# Note: FFmpeg required (install separately via: brew install ffmpeg)
pydub==0.25.1
audioop-lts

# AI API - Google Gemini (FREE)
google-generativeai>=0.3.2
# Note: This package is deprecated, consider migrating to google-genai in future

# ReazonSpeech (K2/ONNX version for Mac)
sherpa_onnx
git+https://github.com/reazon-research/reazonspeech.git#subdirectory=pkg/k2-asr

